requirements.txt
django>=6.0
python-telegram-bot>=20.0
python-dotenv>=1.0
requests>=2.31
